// src/app/interfaces/sales/Customer.ts
export interface ICustomer {
  id: number;
  customerName: string;
  name: string;
  address: string;
  gstNumber: string;
  phoneNumber: string;
  countryId: number;
  countryName: string;
  cityId: number;
  cityName: string;
  stateId: number;
  stateName: string;
  email: string;
  creditPeriod: number;
  creditLimit: number;
  openingBalance: number;
}